+++ cmd/flagship/main.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/OverlayFox/VRC-Stream-Haven/src/geo"
 	"github.com/OverlayFox/VRC-Stream-Haven/src/haven"
-	"github.com/OverlayFox/VRC-Stream-Haven/src/protocols/rtsp"
+	"github.com/OverlayFox/VRC-Stream-Haven/src/protocols/hls"
 	"github.com/OverlayFox/VRC-Stream-Haven/src/protocols/srt"
 )
 
@@ -37,17 +37,17 @@ func main() {
 	}
 	locator, err := geo.NewLocator(ctx, logger, geoConf)
 	if err != nil {
-		logger.Fatal().Err(err).Msg("Failed to create geo locator")
+		logger.Panic().Err(err).Msg("Failed to create geo locator")
 		return
 	}
 
 	haven, err := haven.NewHaven(ctx, logger, locator, "thisisaverysecurepassphrase", "test")
 	if err != nil {
-		logger.Fatal().Err(err).Msg("Failed to create haven")
+		logger.Panic().Err(err).Msg("Failed to create haven")
 		return
 	}
 
-	rtspConf := rtsp.Config{
+	hlsConf := hls.Config{
 		Port:           8554,
 		Address:        "0.0.0.0",
 		Passphrase:     "thisisaverysecurepassphrase",
@@ -55,12 +55,12 @@ func main() {
 		WriteTimeout:   10 * time.Second,
 		WriteQueueSize: 8192,
 	}
-	rtspServer, err := rtsp.New(ctx, logger, rtspConf, haven, locator)
+	hlsServer, err := hls.New(ctx, logger, hlsConf, haven, locator)
 	if err != nil {
-		logger.Fatal().Err(err).Msg("Failed to create RTSP server")
+		logger.Panic().Err(err).Msg("Failed to create HLS server")
 		return
 	}
-	rtspServer.Start()
+	hlsServer.Start()
 
 	srtConfig := srt.Config{
 		Address: "0.0.0.0",
@@ -68,7 +68,7 @@ func main() {
 	}
 	srtServer, err := srt.New(ctx, logger, srtConfig, haven, locator)
 	if err != nil {
-		logger.Fatal().Err(err).Msg("Failed to create SRT server")
+		logger.Panic().Err(err).Msg("Failed to create SRT server")
 		return
 	}
 	srtServer.Start()